feat: Split out the BBCode extension so it can be independently imported
